Mostly work on Example2.R
But also tidied up getSeasonal() in veg-processing.R and checked multi-threading in Example2.  This works fine with some performance gains and now performance loss for using the parallel machinery with only one thread.  However, I have realised that doParallel and foreach packages are not needed for RVCTools.  Instead they should be suggested packages and used in the example scripts. This is not done yet.