Introduced VegObject class to superceed VegSpatial and VegTemporal and a function getVegObject to build these objects.  It also therefore has the possible to store veg data which is not temporally or spatially averaged, or data which is both spatially and temporally averaged, both of which could conceivably be useful.
This function is pretty powerful, it handles cropping, averaging, reading averaged files from disk and using the full .out table from the VegRun object if it is available.  It is basically a combination of getVegSpatial and getVegTemporal, but more cleverly done.

The examples haven't been updated yet, and I haven't removed anything, so nothing should break.  We need to consider if we keep getVegSpatial and getVegTemporal as convenienent wrappers, but I think for sure we should phase out the VegSpatial and VegTemporal objects.