Implemented selectGridcells() call in getField_GUESS(), also tweaked getField() accordingly().
I can now see that it would make the most sense to change the 'getField_FORMAT()' functions to actually returns a Field.  That will be potentially more memory efficient (avoiding a data.table being stored as an element in the return list *and* as a data.table itself), as well as being cleaner and more maintainable.  It seems completely do-able, but will require some time to implement for all Formats,