Many things.  Firstly bugs caught by unit tests that somehow didn't get noticed in previous commit.
The code for subsetting Fields for plotting has been factored out of plotSpatial() into trimFieldsForPlotting() in preparation for re-use in plotSubannual()

Function plotSubannual() now exists to replace plot Seasonal() but is not yet fully developed.

In selectGridcells, the @spatial.extent now stores the gridcells or shapefile that was used.

Alos, any getField_XXXX() functions must now return the spatial.exent.id as "Full" if no cropping was done.