Added the full STAInfo when writing netCDF files.
This isn't necessarily all read yet, but it is a nice and consistent way of doing things, potentially for saving aggregated data.
*But* I have just noticed that writeNetCDF() doesn't handle year-aggregated data.
I think it is because it assumes that "Year" must be be present for building a time axis.
In general, reading and writing aggregated data is a larger job, and probably not necessary at tis stage.