Added support for monthly relative time axis.
Also important bug fix for negtive years *and* now use the as.data.table() instead of the reshape2::melt to reform the data from a multi-dimensional array as read from the netCDF file into a data.table.  This is very useful because reshape2 is deprecated so it would be nice to get rid of this dependency.