Fixed text.multiplier argument to plotSubannual()
Previously, the text multiplier was applied to the plot by
plotSubannual() if the argument was not missing (ie if it was explicitly
passed in by the user). This caused problems if the user explicitly
passed in a value of NULL (e.g. from a wrapper function). Checking
instead for a NULL value of this argument captures both cases and solves
this problem.