plotTemporal(): don't ignore the x-axis label argument
Previously, the x.label argument was ignored by the plotTemporal()
function. This commit changes the function to use this argument in an
analogous way to the way in which the y.label argument is used.

The only difference to how the y-axis label is treated is that if a
y-axis label is not supplied by the caller, a default one is generated
by calling makeYAxis(). I'm not sure what would constitute a good
default x-axis label (perhaps simply "Time" or "Date" would work in this
case).